resend token implementation moved to registration service
